Split package in two: extension and library (#132)
* Split package in two: extension and library

* Linter

* Missing clean:all script
