+++ .gitignore
@@ -3,7 +3,6 @@ lib/
 node_modules/
 *.log
 .eslintcache
-.stylelintcache
 *.egg-info/
 .ipynb_checkpoints
 *.tsbuildinfo
