+++ package.json
@@ -1,6 +1,7 @@
 {
-    "name": "@jupyterlite/xeus",
+    "name": "@jupyterlite/xeus-root",
     "version": "2.1.2",
+    "private": true,
     "description": "JupyterLite loader for Xeus kernels",
     "keywords": [
         "jupyter",
@@ -13,110 +14,37 @@
     },
     "license": "BSD-3-Clause",
     "author": "JupyterLite Contributors",
-    "files": [
-        "lib/**/*.{d.ts,eot,gif,html,jpg,js,js.map,json,png,svg,woff2,ttf}",
-        "style/**/*.{css,js,eot,gif,html,jpg,json,png,svg,woff2,ttf}"
+    "workspaces": [
+        "packages/*"
     ],
-    "main": "lib/index.js",
-    "types": "lib/index.d.ts",
-    "style": "style/index.css",
     "repository": {
         "type": "git",
         "url": "https://github.com/jupyterlite/xeus.git"
     },
     "scripts": {
-        "build": "jlpm build:lib && jlpm build:worker && jlpm build:labextension:dev",
-        "build:prod": "jlpm clean && jlpm build:lib:prod && jlpm build:worker:prod  && jlpm build:labextension",
-        "build:worker": "webpack --config worker.webpack.config.js --mode=development",
-        "build:worker:prod": "webpack --config worker.webpack.config.js --mode=production",
-        "build:labextension": "jupyter labextension build .",
-        "build:labextension:dev": "jupyter labextension build --development True .",
-        "build:lib": "tsc --sourceMap",
-        "build:lib:prod": "tsc",
-        "clean": "jlpm clean:lib",
-        "clean:lib": "rimraf lib tsconfig.tsbuildinfo",
-        "clean:lintcache": "rimraf .eslintcache .stylelintcache",
-        "clean:labextension": "rimraf jupyterlite_xeus/labextension jupyterlite_xeus/_version.py",
-        "clean:all": "jlpm clean:lib && jlpm clean:labextension && jlpm clean:lintcache",
+        "build": "lerna run build",
+        "build:prod": "lerna run build:prod",
+        "clean": "lerna run clean",
+        "clean:all": "lerna run clean:all",
         "eslint": "jlpm eslint:check --fix",
         "eslint:check": "eslint . --cache --ext .ts,.tsx",
-        "install:extension": "jlpm build",
-        "lint": "jlpm stylelint && jlpm prettier && jlpm eslint",
-        "lint:check": "jlpm stylelint:check && jlpm prettier:check && jlpm eslint:check",
+        "lint": "jlpm prettier && jlpm eslint",
+        "lint:check": "jlpm prettier:check && jlpm eslint:check",
         "prettier": "jlpm prettier:base --write --list-different",
         "prettier:base": "prettier \"**/*{.ts,.tsx,.js,.jsx,.css,.json,.md}\"",
         "prettier:check": "jlpm prettier:base --check",
-        "stylelint": "jlpm stylelint:check --fix",
-        "stylelint:check": "stylelint --cache \"style/**/*.css\"",
-        "watch": "run-p watch:src watch:labextension",
-        "watch:src": "tsc -w --sourceMap",
-        "watch:labextension": "jupyter labextension watch ."
-    },
-    "dependencies": {
-        "@jupyterlab/coreutils": "^6",
-        "@jupyterlab/services": "^7",
-        "@jupyterlite/contents": "^0.2.0 || ^0.3.0 || ^0.4.0",
-        "@jupyterlite/kernel": "^0.2.0 || ^0.3.0 || ^0.4.0",
-        "@jupyterlite/server": "^0.2.0 || ^0.3.0 || ^0.4.0",
-        "@lumino/coreutils": "^2",
-        "@lumino/signaling": "^2",
-        "coincident": "^1.2.3",
-        "comlink": "^4.4.1"
+        "watch": "lerna run watch"
     },
     "devDependencies": {
-        "@jupyterlab/builder": "^4.1.0",
-        "@types/json-schema": "^7.0.11",
-        "@types/react": "^18.0.26",
-        "@types/react-addons-linked-state-mixin": "^0.14.22",
         "@typescript-eslint/eslint-plugin": "^6.1.0",
         "@typescript-eslint/parser": "^6.1.0",
-        "css-loader": "^6.7.1",
         "eslint": "^8.36.0",
         "eslint-config-prettier": "^8.8.0",
         "eslint-plugin-prettier": "^5.0.0",
+        "lerna": "^8.1.9",
         "npm-run-all": "^4.1.5",
         "prettier": "^3.0.0",
-        "rimraf": "^5.0.1",
-        "source-map-loader": "^1.0.2",
-        "style-loader": "^3.3.1",
-        "stylelint": "^15.10.1",
-        "stylelint-config-recommended": "^13.0.0",
-        "stylelint-config-standard": "^34.0.0",
-        "stylelint-csstree-validator": "^3.0.0",
-        "stylelint-prettier": "^4.0.0",
-        "ts-loader": "^9.2.6",
-        "typescript": "~5.0.2",
-        "webpack": "^5.87.0",
-        "yjs": "^13.5.0"
-    },
-    "sideEffects": [
-        "style/*.css",
-        "style/index.js"
-    ],
-    "styleModule": "style/index.js",
-    "publishConfig": {
-        "access": "public"
-    },
-    "jupyterlab": {
-        "extension": true,
-        "outputDir": "jupyterlite_xeus/labextension",
-        "sharedPackages": {
-            "@jupyterlite/kernel": {
-                "bundled": false,
-                "singleton": true
-            },
-            "@jupyterlite/server": {
-                "bundled": false,
-                "singleton": true
-            },
-            "@jupyterlite/contents": {
-                "bundled": false,
-                "singleton": true
-            }
-        }
-    },
-    "jupyterlite": {
-        "liteExtension": true
+        "rimraf": "^5.0.1"
     },
     "eslintIgnore": [
         "node_modules",
@@ -194,22 +122,5 @@
                 }
             }
         ]
-    },
-    "stylelint": {
-        "extends": [
-            "stylelint-config-recommended",
-            "stylelint-config-standard",
-            "stylelint-prettier/recommended"
-        ],
-        "plugins": [
-            "stylelint-csstree-validator"
-        ],
-        "rules": {
-            "csstree/validator": true,
-            "property-no-vendor-prefix": null,
-            "selector-class-pattern": "^([a-z][A-z\\d]*)(-[A-z\\d]+)*$",
-            "selector-no-vendor-prefix": null,
-            "value-no-vendor-prefix": null
-        }
     }
 }
