+++ packages/xeus/package.json
@@ -1,0 +1,65 @@
+{
+    "name": "@jupyterlite/xeus",
+    "version": "2.1.2",
+    "description": "JupyterLite Xeus kernels",
+    "keywords": [
+        "jupyter",
+        "jupyterlab",
+        "jupyterlab-extension"
+    ],
+    "homepage": "https://github.com/jupyterlite/xeus",
+    "bugs": {
+        "url": "https://github.com/jupyterlite/xeus/issues"
+    },
+    "license": "BSD-3-Clause",
+    "author": "JupyterLite Contributors",
+    "files": [
+        "lib/**/*.{d.ts,eot,gif,html,jpg,js,js.map,json,png,svg,woff2,ttf}"
+    ],
+    "main": "lib/index.js",
+    "types": "lib/index.d.ts",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/jupyterlite/xeus.git"
+    },
+    "scripts": {
+        "build": "jlpm build:lib && jlpm build:worker",
+        "build:prod": "jlpm clean && jlpm build:lib:prod && jlpm build:worker:prod",
+        "build:worker": "webpack --config worker.webpack.config.js --mode=development",
+        "build:worker:prod": "webpack --config worker.webpack.config.js --mode=production",
+        "build:lib": "tsc --sourceMap",
+        "build:lib:prod": "tsc",
+        "clean": "jlpm clean:lib",
+        "clean:lib": "rimraf lib tsconfig.tsbuildinfo",
+        "clean:all": "jlpm clean:lib",
+        "watch": "run-p watch:src",
+        "watch:src": "tsc -w --sourceMap"
+    },
+    "dependencies": {
+        "@jupyterlab/coreutils": "^6",
+        "@jupyterlab/services": "^7",
+        "@jupyterlite/contents": "^0.2.0 || ^0.3.0 || ^0.4.0",
+        "@jupyterlite/kernel": "^0.2.0 || ^0.3.0 || ^0.4.0",
+        "@jupyterlite/server": "^0.2.0 || ^0.3.0 || ^0.4.0",
+        "@lumino/coreutils": "^2",
+        "@lumino/signaling": "^2",
+        "coincident": "^1.2.3",
+        "comlink": "^4.4.1"
+    },
+    "devDependencies": {
+        "@jupyterlab/builder": "^4.1.0",
+        "@types/json-schema": "^7.0.11",
+        "@types/react": "^18.0.26",
+        "@types/react-addons-linked-state-mixin": "^0.14.22",
+        "npm-run-all": "^4.1.5",
+        "rimraf": "^5.0.1",
+        "source-map-loader": "^1.0.2",
+        "ts-loader": "^9.2.6",
+        "typescript": "~5.0.2",
+        "webpack": "^5.87.0",
+        "yjs": "^13.5.0"
+    },
+    "publishConfig": {
+        "access": "public"
+    }
+}
