+++ packages/xeus/src/index.ts
@@ -1,0 +1,5 @@
+export * from './coincident.worker';
+export * from './comlink.worker';
+export * from './interfaces';
+export * from './web_worker_kernel';
+export * from './worker';
