+++ packages/xeus/worker.webpack.config.js
