--- style/index.js
@@ -1,1 +1,0 @@
-import './base.css';
